# openapi.json
{
  "components": {
    "schemas": {
...
      "BetaPoint": {
        "properties": {
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "date": {
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "beta"
        ],
        "title": "BetaPoint",
        "type": "object"
      },
...
      "FxEvidenceOut": {
        "description": "Shared public evidence shape for every base-currency analysis.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "base_currency": {
            "title": "Base Currency",
            "type": "string"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "missing_currencies": {
            "items": {
              "type": "string"
            },
            "title": "Missing Currencies",
            "type": "array"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "enum": [
              "identity",
              "converted",
              "incomplete"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "base_currency",
          "source",
          "as_of",
          "fetched_at",
          "missing_currencies",
          "note"
        ],
        "title": "FxEvidenceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Histogram": {
        "properties": {
          "bin_edges": {
            "items": {
              "type": "number"
            },
            "title": "Bin Edges",
            "type": "array"
          },
          "counts": {
            "items": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "array"
          }
        },
        "required": [
          "bin_edges",
          "counts"
        ],
        "title": "Histogram",
        "type": "object"
      },
...
      "MonteCarloRequest": {
        "properties": {
          "horizon": {
            "default": 252,
            "maximum": 2520.0,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          },
          "n_paths": {
            "default": 10000,
            "maximum": 200000.0,
            "minimum": 1.0,
            "title": "N Paths",
            "type": "integer"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "MonteCarloRequest",
        "type": "object"
      },
      "MonteCarloResponse": {
        "properties": {
          "es_975": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Es 975"
          },
          "fx": {
            "$ref": "#/components/schemas/FxEvidenceOut"
          },
          "histogram": {
            "$ref": "#/components/schemas/Histogram"
          },
          "horizon": {
            "title": "Horizon",
            "type": "integer"
          },
          "n_nonfinite": {
            "title": "N Nonfinite",
            "type": "integer"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "p5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P5"
          },
          "p50": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50"
          },
          "p95": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "fx",
          "symbol",
          "horizon",
          "n_paths",
          "histogram",
          "p5",
          "p50",
          "p95",
          "es_975",
          "n_nonfinite"
        ],
        "title": "MonteCarloResponse",
        "type": "object"
      },
...
      "RiskResponse": {
        "properties": {
          "alpha_annualized": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha Annualized"
          },
          "alpha_note": {
            "title": "Alpha Note",
            "type": "string"
          },
          "ann_vol": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ann Vol"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "beta_series": {
            "items": {
              "$ref": "#/components/schemas/BetaPoint"
            },
            "title": "Beta Series",
            "type": "array"
          },
          "cagr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cagr"
          },
          "drag_approx": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drag Approx"
          },
          "drag_exact": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drag Exact"
          },
          "drag_note": {
            "title": "Drag Note",
            "type": "string"
          },
          "es_975": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Es 975"
          },
          "fx": {
            "$ref": "#/components/schemas/FxEvidenceOut"
          },
          "mean_arith_annual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Arith Annual"
          },
          "n_obs": {
            "title": "N Obs",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "integer"
          },
          "years": {
            "title": "Years",
            "type": "integer"
          }
        },
        "required": [
          "fx",
          "symbol",
          "benchmark",
          "window",
          "years",
          "n_obs",
          "beta_series",
          "alpha_annualized",
          "alpha_note",
          "es_975",
          "ann_vol",
          "mean_arith_annual",
          "cagr",
          "drag_exact",
          "drag_approx",
          "drag_note",
          "as_of"
        ],
        "title": "RiskResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "QuantMind API",
    "version": "0.4.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/risk/montecarlo": {
      "post": {
        "operationId": "montecarlo_api_risk_montecarlo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonteCarloRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonteCarloResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Montecarlo"
      }
    },
    "/api/risk/{symbol}": {
      "get": {
        "operationId": "risk_api_risk__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 60,
              "maximum": 756,
              "minimum": 5,
              "title": "Window",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "years",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 25,
              "minimum": 1,
              "title": "Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk"
      }
    },
    "/api/risk/{symbol}/regression": {
      "get": {
        "operationId": "regression_api_risk__symbol__regression_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "description": "comma-separated factor names, e.g. SPY,MTUM,US10Y",
            "in": "query",
            "name": "factors",
            "required": true,
            "schema": {
              "description": "comma-separated factor names, e.g. SPY,MTUM,US10Y",
              "title": "Factors",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2520,
                  "minimum": 20,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window"
            }
          },
          {
            "in": "query",
            "name": "years",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 25,
              "minimum": 1,
              "title": "Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regression"
      }
    },
...
  }
}